Preserve original sort position if order is not provided
Despite documentation suggesting otherwise, returning a 0
in a sort comparator does not actually preserve the item's
position in the original array's sort order. This means that if
you didn't provide any manual ordering of your items, they'd
appear in random order as opposed to the order they appear
in your CSS.

This custom sort function keeps track of the item's original
index and uses that if no order is found.